refactor(directive): extract file directive + listener helpers (DIR-F7)
Move the file-input directive variant + its dependencies out of
`directive.ts` (~155 LOC for the file cluster, ~30 LOC for the
listener helpers) into two focused modules:

  - `directive-listeners.ts` (~85 LOC) — shared by every directive
    variant. Owns the per-element `listenersKey` symbol slot, the
    `TrackedListener` tuple, and the `addTrackedListener` /
    `removeTrackedListeners` / `noteInteraction` helpers. `add` is
    renamed from `addEventListener` so the file's tracked path is
    distinct in greps from the DOM's built-in `el.addEventListener`.
    `noteInteraction` carries a leaner local `isRegisterValueLike`
    type guard rather than reaching back across the module boundary
    for the public `isRegisterValue` export — avoids a circular
    import and narrows to just the `markInteracted` method it needs.

  - `directive-file.ts` (~175 LOC) — owns `isBlankFileValue`,
    `readFilesFromInput`, the `maybeWarnPersistedFile` dev warn
    book + dedupe map, the `fileScopeKey` symbol, and the
    `vRegisterFile` directive object itself. `vRegisterFile` keeps
    its full `created` / `beforeUpdate` / `beforeUnmount` lifecycle
    semantics — blank-shape seeding, the storage→DOM blank-resync
    watcher in its own `effectScope`, and the per-element teardown.

`directive.ts` re-imports the listener helpers and `vRegisterFile`,
trimming the host file by ~190 LOC. Bulk-rename of `addEventListener
(el|input,` to `addTrackedListener` across the 10 directive variant
call sites. Drops the now-unused `effectScope` / `watch` / `warn`-
file-warn / `PersistOptInRegistry` imports from `directive.ts`.

Behavior-neutral: file directive tests + the full 3547-test suite
stay green unchanged.
