+++ src/runtime/core/directive.ts
@@ -25,7 +25,7 @@ import {
   looseToNumber,
 } from './vue-shared-shim'
 import type { DirectiveBinding, DirectiveHook, ObjectDirective, VNode } from 'vue'
-import { effectScope, isRef, nextTick, warn, watch } from 'vue'
+import { isRef, nextTick, warn } from 'vue'
 import { REGISTER_OWNER_MARKER } from '../composables/use-register'
 import { __DEV__ } from './dev'
 import {
@@ -36,6 +36,8 @@ import {
   teardownAria,
   type AriaCarrier,
 } from './directive-aria'
+import { vRegisterFile } from './directive-file'
+import { addTrackedListener, noteInteraction, removeTrackedListeners } from './directive-listeners'
 import { INTERACTIVE_TAG_NAMES } from './interactive-tags'
 import type {
   CustomDirectiveRegisterAssignerFn,
@@ -50,7 +52,6 @@ import type {
   WriteMeta,
 } from '../types/types-api'
 import type { PathKey } from './paths'
-import type { PersistOptInRegistry } from './persistence/opt-in-registry'
 import { getOrAssignElementId } from './persistence/opt-in-registry'
 import { allowSensitivePersist } from './persistence/sensitive-names'
 
@@ -78,24 +79,6 @@ import { allowSensitivePersist } from './persistence/sensitive-names'
 export const assignKey: unique symbol = Symbol.for('attaform:assign-key')
 
 /**
- * Per-element bag of listener tuples added by the active directive
- * variant in `created`. `vRegisterDynamic.beforeUnmount` drains the bag
- * so reused elements (KeepAlive, v-show) don't accumulate orphaned
- * handlers across activation cycles.
- */
-const listenersKey: unique symbol = Symbol.for('attaform:directive-listeners')
-
-type TrackedListener = {
-  event: string
-  handler: EventListener
-  // Explicitly `undefined`-able so `exactOptionalPropertyTypes` lets us
-  // stash tuples where the caller didn't pass options.
-  options: EventListenerOptions | undefined
-}
-
-type ListenerCarrier = { [listenersKey]?: TrackedListener[] }
-
-/**
  * Type guard for a `RegisterValue`. Returns `true` when `val` looks
  * like the object returned from `form.register(path)`.
  *
@@ -120,31 +103,6 @@ export function isRegisterValue<Value = unknown>(val: unknown): val is RegisterV
 }
 
 type ComposingTarget = (EventTarget & { composing: boolean }) | null
-function addEventListener(
-  el: Element,
-  event: string,
-  handler: EventListener,
-  options?: EventListenerOptions
-): void {
-  el.addEventListener(event, handler, options)
-  // Stash the tuple on the element so `beforeUnmount` can detach it.
-  // A bare `addEventListener` without tracking would leak across
-  // KeepAlive re-activations where the DOM node is reused.
-  const carrier = el as ListenerCarrier
-  const bag = carrier[listenersKey] ?? []
-  bag.push({ event, handler, options })
-  carrier[listenersKey] = bag
-}
-
-function removeTrackedListeners(el: Element): void {
-  const carrier = el as ListenerCarrier
-  const bag = carrier[listenersKey]
-  if (bag === undefined) return
-  for (const { event, handler, options } of bag) {
-    el.removeEventListener(event, handler, options)
-  }
-  delete carrier[listenersKey]
-}
 
 /**
  * Write the directive-private `lastTypedForm` ref. Lives on the
@@ -633,16 +591,6 @@ function setAssignFunction(
 
 // We are exporting the v-model runtime directly as vnode hooks so that it can
 // be tree-shaken in case v-model is never used.
-// First genuine user-input event flips the field's sticky `interacted`
-// bit — the signal `defaultDisplayState` reads to keep a clean
-// tab-through quiet while still engaging validation the moment the user
-// edits. Routed only through these DOM listeners, so hydration and
-// programmatic setValue never trip it. Idempotent and store-guarded on
-// the RegisterValue side.
-function noteInteraction(value: unknown): void {
-  if (isRegisterValue(value)) value.markInteracted()
-}
-
 const vRegisterText: RegisterTextCustomDirective = {
   created(el, { value, modifiers: { lazy, trim, number } }, vnode) {
     const castToNumber = number === true || vnode.props?.['type'] === 'number'
@@ -650,7 +598,7 @@ const vRegisterText: RegisterTextCustomDirective = {
       value.registerElement(el)
       setAssignFunction(el, vnode, value)
     }
-    addEventListener(el, lazy === true ? 'change' : 'input', (e) => {
+    addTrackedListener(el, lazy === true ? 'change' : 'input', (e) => {
       // Bail if this listener was attached on a non-supported root
       // (a `<label>` / `<div>` etc.) AND the assigner is the default.
       // The bubbled-write bug fires here without this guard: a
@@ -819,7 +767,7 @@ const vRegisterText: RegisterTextCustomDirective = {
       }
     })
     if (trim === true || castToNumber) {
-      addEventListener(el, 'change', () => {
+      addTrackedListener(el, 'change', () => {
         if (shouldBailListener(el)) return
         // Mirror Vue's `castValue(el.value, trim, castToNumber)` so the
         // visible DOM normalizes after blur for both modifiers — without
@@ -872,13 +820,13 @@ const vRegisterText: RegisterTextCustomDirective = {
       })
     }
     if (lazy !== true) {
-      addEventListener(el, 'compositionstart', onCompositionStart)
-      addEventListener(el, 'compositionend', onCompositionEnd)
+      addTrackedListener(el, 'compositionstart', onCompositionStart)
+      addTrackedListener(el, 'compositionend', onCompositionEnd)
       // Safari < 10.2 & UIWebView doesn't fire compositionend when
       // switching focus before confirming composition choice
       // this also fixes the issue where some browsers e.g. iOS Chrome
       // fires "change" instead of "input" on autocomplete.
-      addEventListener(el, 'change', onCompositionEnd)
+      addTrackedListener(el, 'change', onCompositionEnd)
     }
     // `.number` × text input — block non-numeric characters at the
     // DOM layer so `el.value` never holds garbage. Native
@@ -895,7 +843,7 @@ const vRegisterText: RegisterTextCustomDirective = {
     // normally and the directive's `compositionend` handler catches
     // the final value.
     if (number === true && vnode.props?.['type'] !== 'number') {
-      addEventListener(el, 'beforeinput', (e) => {
+      addTrackedListener(el, 'beforeinput', (e) => {
         const ev = e as InputEvent
         if (
           ev.inputType !== 'insertText' &&
@@ -982,7 +930,7 @@ const vRegisterCheckbox: RegisterCheckboxCustomDirective = {
 
     value.registerElement(el)
     setAssignFunction(el, vnode, value)
-    addEventListener(el, 'change', () => {
+    addTrackedListener(el, 'change', () => {
       if (shouldBailListener(el)) return
       noteInteraction(value)
       const modelValue = value.innerRef.value ?? []
@@ -1137,7 +1085,7 @@ const vRegisterRadio: RegisterRadioCustomDirective = {
 
     value.registerElement(el)
     setAssignFunction(el, vnode, value)
-    addEventListener(el, 'change', () => {
+    addTrackedListener(el, 'change', () => {
       if (shouldBailListener(el)) return
       noteInteraction(value)
       el[assignKey]?.(getValue(el))
@@ -1200,7 +1148,7 @@ const vRegisterSelect: RegisterSelectCustomDirective = {
 
     value.registerElement(el)
     const isSetModel = isSet(value.innerRef.value)
-    addEventListener(el, 'change', () => {
+    addTrackedListener(el, 'change', () => {
       if (shouldBailListener(el)) return
       noteInteraction(value)
       const selectedVal = Array.prototype.filter
@@ -1582,164 +1530,6 @@ const vRegisterDynamic: RegisterModelDynamicCustomDirective = {
     // SSR path; under compiled SSR an authored aria attribute can't be
     // detected here, and the client directive reconciles it on hydration.
     return getSSRAriaProps(rv, (vnode as VNode | null) ?? null)
-  },
-}
-
-// True for any value the file directive treats as "no file selected":
-// `null`, `undefined`, the empty array (multi-input cleared), and an
-// empty `FileList`. Strict equality short-circuits the common cases;
-// the FileList check covers the case where a host wrote the live DOM
-// FileList back into storage (rare, but cheap to handle).
-function isBlankFileValue(value: unknown): boolean {
-  if (value === null || value === undefined) return true
-  if (Array.isArray(value) && value.length === 0) return true
-  if (typeof FileList !== 'undefined' && value instanceof FileList && value.length === 0)
-    return true
-  return false
-}
-
-// Read the current selection off a file input and reshape to the
-// directive's canonical storage form: `File[]` when the element has the
-// `multiple` attribute, `File | null` otherwise. `el.files` is `null`
-// on programmatically-detached inputs and the FileList is empty when
-// the user picked nothing — both collapse to the blank shape.
-function readFilesFromInput(el: HTMLInputElement): File[] | File | null {
-  const files = el.files
-  if (el.multiple) {
-    return files === null ? [] : Array.from(files)
-  }
-  if (files === null || files.length === 0) return null
-  return files.item(0)
-}
-
-// Per-form dedupe for the persisted-file-input dev warning. Keyed on
-// the form's `PersistOptInRegistry` (every form has its own instance)
-// so the warning fires once per (form, path) — a single noisy hint
-// during development rather than per-mount, per-keystroke noise.
-const warnedPersistedFileForms: WeakMap<PersistOptInRegistry, Set<PathKey>> | null = __DEV__
-  ? new WeakMap<PersistOptInRegistry, Set<PathKey>>()
-  : null
-
-function maybeWarnPersistedFile(value: RegisterValue): void {
-  if (!__DEV__ || warnedPersistedFileForms === null) return
-  if (value.persist !== true) return
-  let warnedPaths = warnedPersistedFileForms.get(value.persistOptIns)
-  if (warnedPaths === undefined) {
-    warnedPaths = new Set<PathKey>()
-    warnedPersistedFileForms.set(value.persistOptIns, warnedPaths)
-  }
-  if (warnedPaths.has(value.path)) return
-  warnedPaths.add(value.path)
-  warn(
-    `[attaform] register('${value.path}', { persist: true }) on <input type="file"> — ` +
-      `files can't ride a refresh (browsers block programmatic writes to ` +
-      `<input type="file">), so this path won't be saved. For long-lived ` +
-      `flows, upload on selection and persist the resulting URL or ID in a ` +
-      `sibling string field.`
-  )
-}
-
-// Real `v-register` variant for `<input type="file">`. Reads
-// `event.target.files` into form state as `File | null` (single) or
-// `File[]` (multiple). Storage is the canonical blank shape (`null` /
-// `[]`) when no file is selected, with the path marked in
-// `blankPaths` so the friendly "No value supplied" error surfaces
-// through `derivedBlankErrors` on required-file fields — same channel
-// as required numbers / bigints.
-//
-// The persistence carve-out lives in `syncPersistOptIn`: file paths
-// never enter `persistOptIns`, never serialize, never rehydrate. The
-// `beforeUpdate` hook keeps the DOM in lockstep with storage by
-// clearing `el.value` when storage transitions to blank — the only
-// programmatic write browsers permit on file inputs.
-// Symbol slot for the per-element effect-scope teardown function. The
-// blank-resync watcher inside `created` runs in its own scope so we
-// can stop it on `beforeUnmount` without depending on the surrounding
-// component still being alive.
-const fileScopeKey: unique symbol = Symbol.for('attaform:file-scope')
-type FileScopeCarrier = { [fileScopeKey]?: () => void }
-
-const vRegisterFile: RegisterModelDynamicCustomDirective = {
-  created(el, { value }) {
-    if (!isRegisterValue(value)) return
-    // `resolveDynamicModel` routes here only when `el.tagName === 'INPUT'`
-    // and `el.type === 'file'`. The variant union type widens to include
-    // select/textarea, so narrow once per hook.
-    const input = el as HTMLInputElement
-    value.registerElement(input)
-    maybeWarnPersistedFile(value)
-
-    // Seed the blank-path channel on register. Storage shape gets
-    // canonicalised to `null` / `[]` whenever the consumer's default
-    // is loosely blank (e.g. `undefined` for a non-nullable
-    // `z.file()` schema), so reads return a uniform shape regardless
-    // of how the user expressed "optional file" in their schema.
-    const currentRaw = value.innerRef.value
-    if (isBlankFileValue(currentRaw)) {
-      const blankShape: File[] | null = input.multiple ? [] : null
-      value.setValueWithInternalPath(blankShape, { blank: true })
-    }
-
-    addEventListener(input, 'change', () => {
-      noteInteraction(value)
-      const next = readFilesFromInput(input)
-      const blank = isBlankFileValue(next)
-      value.setValueWithInternalPath(next, blank ? { blank: true } : undefined)
-    })
-
-    // Watch storage for programmatic transitions to the blank shape
-    // (`form.clear(path)` / `form.reset()` / hydrate). Re-mark the
-    // path blank and clear the DOM input. `beforeUpdate` covers the
-    // common parent-re-render case; this watcher catches storage
-    // mutations that don't trigger a parent re-render. Runs in its
-    // own effect scope so we can stop it from `beforeUnmount`
-    // independent of the surrounding component.
-    const scope = effectScope(true)
-    scope.run(() => {
-      watch(
-        value.innerRef,
-        (next) => {
-          if (!isBlankFileValue(next)) return
-          value.setValueWithInternalPath(next, { blank: true })
-          if (input.value !== '') input.value = ''
-        },
-        { flush: 'post' }
-      )
-    })
-    ;(input as FileScopeCarrier)[fileScopeKey] = (): void => scope.stop()
-  },
-  beforeUpdate(el, { value }) {
-    if (!isRegisterValue(value)) return
-    const input = el as HTMLInputElement
-    // Storage → DOM + blankPaths sync. Two responsibilities:
-    //
-    //   1. Clear the DOM input when storage went blank
-    //      (`form.clear(path)` / `form.reset()` / hydrate). `el.value
-    //      = ''` is the one programmatic mutation browsers allow on
-    //      `<input type="file">`.
-    //
-    //   2. Re-mark the path blank in the store so `derivedBlankErrors`
-    //      keeps firing the friendly "No value supplied" message after
-    //      programmatic clears. `form.clear` writes the schema's empty
-    //      value (`null`) but doesn't propagate `meta.blank: true`, so
-    //      the path would otherwise drift out of `blankPaths`. The
-    //      store's `Set.add` is idempotent, and identity-equal writes
-    //      don't trigger re-renders — safe to call on every update.
-    const currentRaw = value.innerRef.value
-    if (isBlankFileValue(currentRaw)) {
-      value.setValueWithInternalPath(currentRaw, { blank: true })
-      if (input.value !== '') input.value = ''
-    }
-  },
-  beforeUnmount(el, { value }) {
-    removeTrackedListeners(el)
-    const stop = (el as FileScopeCarrier)[fileScopeKey]
-    if (stop !== undefined) {
-      stop()
-      delete (el as FileScopeCarrier)[fileScopeKey]
-    }
-    if (!isRegisterValue(value)) return
-    value.deregisterElement(el)
   },
 }
